% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p3theme.R
\name{add_p3theme}
\alias{add_p3theme}
\title{Add p3 theme}
\usage{
add_p3theme(ggplot_obj)
}
\arguments{
\item{ggplot_obj}{A ggplot object}
}
\value{
A ggplot object with p3 ggthemes
}
\description{
Add p3 theme
}
