% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p3theme.R
\name{add_p3theme}
\alias{add_p3theme}
\title{Add p3 theme}
\usage{
add_p3theme(ggplot_obj, target = "fill")
}
\arguments{
\item{ggplot_obj}{A ggplot object}

\item{target}{The aesthetic to which the p3 ggtheme will be applied. The default value is "fill." The other option is "color."}
}
\value{
A ggplot object with p3 ggtheme
}
\description{
Add p3 theme
}
